# By Igor Yudintsev
# Via Igor Yudintsev (1) and Michael Lukashov (1)
* https://github.com/michaellukashov/filecopyex3:
  removed check for NULL on new/delete
  DescList load/save methods return bool
  fix for some static analyzer warnings
  flags, attributes, selection indices are unsigned
  more INVALID_FILE_ATTRIBUTES
  INVALID_FILE_ATTRIBUTES

Conflicts:
	src/Engine.cpp

+++ src/FarMenu.cpp
@@ -70,7 +70,7 @@ void FarMenu::SetBottom(const String & v)
   Bottom = v;
 }
 
-void FarMenu::SetFlags(int f)
+void FarMenu::SetFlags(DWORD f)
 {
   Flags = f;
 }
@@ -80,7 +80,7 @@ void FarMenu::SetHelpTopic(const String & v)
   HelpTopic = v;
 }
 
-void FarMenu::SetSelection(int n)
+void FarMenu::SetSelection(size_t n)
 {
   Selection = n;
 }
